불필요한 기능 제거(글귀 게시판 디테일 페이지 보기 기능 제거

+++ BookieProject/src/main/java/com/project/bookie/service/ClubBoardService.java
@@ -15,7 +15,7 @@ public class ClubBoardService {
 	ClubBoardMapper mapper;
 
 
-	public Board getBoardByBoardById(int boardId) {
+	public Board getBoardByBoardById(long boardId) {
 		Board board = mapper.getBoardByBoardIdWithComment(boardId);
 		return board;
 	}
